use JsonMapperTestUtil in serialisation tests, utilities moved to own package,
