+++ sbt_common/display/src/test/scala/uk/ac/wellcome/display/models/DisplayAbstractConceptSerialisationTest.scala
@@ -1,12 +1,8 @@
 package uk.ac.wellcome.display.models
 
 import org.scalatest.FunSpec
-import uk.ac.wellcome.models.work.internal.{
-  AbstractConcept,
-  Concept,
-  Period,
-  Place
-}
+import uk.ac.wellcome.display.test.util.JsonMapperTestUtil
+import uk.ac.wellcome.models.work.internal.{AbstractConcept, Concept, Period, Place}
 
 class DisplayAbstractConceptSerialisationTest
     extends FunSpec
