+++ sbt_common/display/src/test/scala/uk/ac/wellcome/display/models/v2/DisplayLocationsV2SerialisationTest.scala
@@ -1,17 +1,9 @@
 package uk.ac.wellcome.display.models.v2
 
 import org.scalatest.FunSpec
-import uk.ac.wellcome.display.models.{
-  DisplaySerialisationTestBase,
-  JsonMapperTestUtil,
-  WorksIncludes,
-  WorksUtil
-}
-import uk.ac.wellcome.models.work.internal.{
-  IdentifiedItem,
-  IdentifiedWork,
-  PhysicalLocation
-}
+import uk.ac.wellcome.display.models.{DisplaySerialisationTestBase, WorksIncludes}
+import uk.ac.wellcome.display.test.util.{JsonMapperTestUtil, WorksUtil}
+import uk.ac.wellcome.models.work.internal.{IdentifiedItem, IdentifiedWork, PhysicalLocation}
 
 class DisplayLocationsV2SerialisationTest
     extends FunSpec
