+++ sbt_common/display/src/test/scala/uk/ac/wellcome/display/models/v2/DisplayPublicationDateV2SerialisationTest.scala
@@ -1,21 +1,15 @@
 package uk.ac.wellcome.display.models.v2
 
-import com.fasterxml.jackson.databind.ObjectMapper
-import com.google.inject.Guice
 import org.scalatest.FunSpec
-import uk.ac.wellcome.display.models.{DisplaySerialisationTestBase, WorksUtil}
-import uk.ac.wellcome.display.modules.DisplayJacksonModule
+import uk.ac.wellcome.display.models.DisplaySerialisationTestBase
+import uk.ac.wellcome.display.test.util.{JsonMapperTestUtil, WorksUtil}
 import uk.ac.wellcome.models.work.internal.{IdentifiedWork, Period}
-import uk.ac.wellcome.test.utils.JsonTestUtil
 
 class DisplayPublicationDateV2SerialisationTest
     extends FunSpec
     with DisplaySerialisationTestBase
-    with JsonTestUtil
+    with JsonMapperTestUtil
     with WorksUtil {
-  val injector = Guice.createInjector(DisplayJacksonModule)
-
-  val objectMapper = injector.getInstance(classOf[ObjectMapper])
 
   it("omits the publicationDate field if it is empty") {
     val work = IdentifiedWork(
