+++ sbt_common/display/src/test/scala/uk/ac/wellcome/display/models/v2/DisplayWorkV2SerialisationTest.scala
@@ -1,25 +1,15 @@
 package uk.ac.wellcome.display.models.v2
 
-import com.fasterxml.jackson.databind.ObjectMapper
-import com.google.inject.Guice
 import org.scalatest.FunSpec
-import uk.ac.wellcome.display.models.{
-  DisplaySerialisationTestBase,
-  WorksIncludes,
-  WorksUtil
-}
-import uk.ac.wellcome.display.modules.DisplayJacksonModule
+import uk.ac.wellcome.display.models.{DisplaySerialisationTestBase, WorksIncludes}
+import uk.ac.wellcome.display.test.util.{JsonMapperTestUtil, WorksUtil}
 import uk.ac.wellcome.models.work.internal._
-import uk.ac.wellcome.test.utils.JsonTestUtil
 
 class DisplayWorkV2SerialisationTest
     extends FunSpec
     with DisplaySerialisationTestBase
-    with JsonTestUtil
+    with JsonMapperTestUtil
     with WorksUtil {
-  val injector = Guice.createInjector(DisplayJacksonModule)
-
-  val objectMapper = injector.getInstance(classOf[ObjectMapper])
 
   it("serialises a DisplayWorkV2 correctly") {
